% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCommentFunction.R
\name{fetchComments}
\alias{fetchComments}
\title{Helper function to fetch Comments from Reddit Post}
\usage{
fetchComments(permalink, headers, maxComments)
}
\arguments{
\item{permalink}{The permalink to the Reddit post.}

\item{headers}{Declares Oauth Authentication.}

\item{maxComments}{Maximum number of comments to fetch.}
}
\value{
A list of comments with details like body, author, and creation date per post.
}
\description{
Helper function to fetch Comments from Reddit Post
}
