% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRequestFunction.R
\name{makeRequest}
\alias{makeRequest}
\title{Helper function that builds Reddit API Request}
\usage{
makeRequest(url, headers, params)
}
\arguments{
\item{url}{A string containing the API endpoint.}

\item{headers}{Declares Oauth Authentication.}

\item{params}{A list of query parameters to include in the request.}
}
\value{
Parsed JSON content from the response.
}
\description{
Helper function that builds Reddit API Request
}
