% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchRedditFunction.R
\name{searchReddit}
\alias{searchReddit}
\title{Helper function to search reddit posts}
\usage{
searchReddit(
  userAgent,
  subreddit = NA,
  query = NA,
  startTime = NA,
  endTime = NA,
  sort = "relevance",
  time = "all",
  type = "query",
  batchSize = 100,
  getComments = FALSE,
  maxComments = 100
)
}
\arguments{
\item{userAgent}{username used for Oauth authentication.}

\item{subreddit}{(Optional) restricts search to subreddit if set to 1.}

\item{query}{The search term or author name, depending on `type`.}

\item{startTime}{(Optional) Start of the search range (POSIXct).}

\item{endTime}{(Optional) End of the search range (POSIXct).}

\item{sort}{Sorting criteria: "relevance", "new", "hot", etc.}

\item{time}{Time range: "all", "day", "week", "year".}

\item{type}{Type of search: "query" or "author".}

\item{batchSize}{Number of posts to fetch per keyword.}

\item{getComments}{Whether to fetch comments for each post.}

\item{maxComments}{Maximum number of comments to fetch per post.}
}
\value{
A list of posts, each with details such as title, body, created time, subreddit, and comments.
}
\description{
Helper function to search reddit posts
}
